hotfix : 팀스코어 계산 로직 수정 및 리마인드 알림 eventPublisher 삭제

+++ src/main/java/com/moing/backend/domain/teamScore/application/service/TeamScoreLogicUseCase.java
@@ -41,7 +41,6 @@ public class TeamScoreLogicUseCase {
     public Long updateTeamScore(Long missionId) {
         Mission mission = missionQueryService.findMissionById(missionId);
         Team team = mission.getTeam();
-
         TeamScore teamScore = teamScoreQueryService.findTeamScoreByTeam(team.getTeamId());
 
         teamScore.updateScore(getScoreByMission(mission));
@@ -62,28 +61,25 @@ public class TeamScoreLogicUseCase {
     }
 
     public Long getScoreByMission(Mission mission) {
-        Long total = totalPeople(mission);
-        Long done = donePeople(mission);
+        float total = totalPeople(mission);
+        float done = donePeople(mission);
 
         if (done == 0) {
             return 0L;
         } else {
-            return (done / total * 100) / 5 ;
+            return (long) ((done / total * 100) / 5);
         }
 
     }
 
-    public Long donePeople(Mission mission) {
-        return Long.valueOf(missionStateQueryService.stateCountByMissionId(mission.getId()));
+    public float donePeople(Mission mission) {
+        return Float.valueOf(missionStateQueryService.stateCountByMissionId(mission.getId()));
     }
 
-    public Long totalPeople(Mission mission) {
-        return Long.valueOf(mission.getTeam().getNumOfMember());
+    public float totalPeople(Mission mission) {
+        return Float.valueOf(mission.getTeam().getNumOfMember());
 
     }
-
-
-
